app-python/requirements.txt
Flask
grpcio
grpcio-tools
Flask-RESTful
python-dotenv
flask_table
confluent-kafka